fix finding the next payable
Previously, we searched backward from the payment made ('Z' record) to find a payable ('A' record) that had not been fully satisfied. If the person was all caught up, the __prev_due_idx() method would fail to return an index to apply the payment to.  With this change, if no PRIOR 'A' records (payable) is open, then we'll look FORWARD until we find a future payable that has not been paid in full.